Add environment method to plugin api
This commit adds an environment method to the AbstractExecFork mirroring the
same method in gradle Exec Tasks. This method allows to add environment
variables to the forked process.

It also changes the way output of the executed process is printed to the gradle
command line by using the logger instead of prinitng it to sysout directly. This
way the progress information printed by gradle doesn't bleed into the process
log.